changed file names to be more accurate
