--- src/Home/Connect.jsx
@@ -1,9 +1,0 @@
-// import React, { useState, useEffect, Component } from 'react';
-
-// const Connect = () => {
-//   return(
-
-//   )
-// }
-
-// export default Connect
