+++ src/Home/HomeMain.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, Component} from 'react';
-import UserInput from './UserInput.jsx';
+import InputRows from './InputRows.jsx';
 // import Connect from 'Connect';
 
 function Home() {
@@ -12,10 +12,10 @@ function Home() {
 
   let fieldsArray = [];
   for (let i = 0; i < fields.length; i++) {
-    fieldsArray.push(<UserInput fields={fields[i]} key={i}/>)
+    fieldsArray.push(<InputRows fields={fields[i]} key={i}/>)
   }
 
-  return(
+  return( //replaces "render"
     <div className="homeContainer">
       <h1>"I Love Aki" - Adi</h1>
       {fieldsArray}
