+++ src/Home/InputRows.jsx
@@ -1,10 +1,18 @@
-import React, { useState, useEffect, Component } from 'react'
+import React, { useState, useEffect, Component } from 'react';
+import UserInput from './UserInput.jsx'
 
 const InputRows = (props) => {
+
+  //['Host', 'Database'],
+  const fieldsArray = [];
+  let keyID = 0
+  props.fields.forEach((el) => {
+    fieldsArray.push(<UserInput field={el} key={++keyID}/>)
+  })
+
   return(
-    <div className="UserInput">
-      {props.field}
-      <input></input>
+    <div className="InputRows">
+      {fieldsArray}
     </div>
   )
 }
