+++ src/styles.scss
@@ -32,7 +32,7 @@ body {
       h1{
         text-align: center;
       }
-      .UserRows {
+      .InputRows {
         display: flex;
         align-items: center;
         justify-content: space-evenly;
